Move unit rectangle test to test_patches
